prepare adhs python packaging
To test:

    $ cd adhs
    $ mkvirtualenv adhs
    (adhs) $ python setup.py sdist
    (adhs) $ pip install dist/adhs-0.1.tar.gz
    ...
    (adhs) $ adhs.py path/to/triples.nt
     * Running on http://127.0.0.1:5000/ (Press CTRL+C to quit)

adhs could now be uploaded to [pypi](http://pypi.python.org/),
so everyone could install it with:

    $ pip install adhs

----

Notes:

It's cleaner to move templates under a package, so a package
has been created. The name "adhs" couldn't serve as package
name, since the executable has the same name and shadows the
package.

So the package name is adhocsparql for now.

Adjusted template loader, so it can access templates from the
package as well (see: jinja2.ChoiceLoader).